backend/pyproject.toml
[build-system]
requires = ["hatchling >= 1.26"]
build-backend = "hatchling.build"

# Build tool configuration
[tool.hatch.build.targets.wheel]
packages = ["."]

# Project metadata
[project]
name = "strumok"
version = "0.1.0"
description = "Co-op electricity meter reporting app"
requires-python = ">=3.14"
authors = [{ name = "natikos" }]
dependencies = [
  "fastapi>=0.116.1",
  "uvicorn[standard]>=0.35.0",
  "pydantic-settings>=2.12",
  "sqlmodel>=0.0.33",
  "passlib[bcrypt]>=1.7.4",
  "bcrypt==4.0.1",
  "psycopg[binary]>=3.3.2",
]

[project.optional-dependencies]
dev = ["ruff>=0.15.1"]
